% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.byComponent.R
\name{assemble.byComponent}
\alias{assemble.byComponent}
\title{Assemble Files Based on Component}
\usage{
assemble.byComponent(SJDScorePlotter.obj, component, SJD_algorithm, group = NA)
}
\arguments{
\item{SJDScorePlotter.obj}{A list outputted by the SJDScorePlotter function}

\item{component}{numer/order of component of interest to print out, i.e. 1 or c(1,2)}

\item{SJD_algorithm}{SJD_algorithm name of SJD algorithm, i.e. concatICA}

\item{group}{group name of the weights group from the SJD_algorithm output, i.e 'Shared.All.13'}
}
\value{
A list of images filtered by component
}
\description{
Assemble individual same-component figures from multiple datasets analyzed and plotted by SJD for cross comparison
}
\examples{

library(ggplot2)

data(NeuroGenesis4.afterWrap)
data(NeuroGenesis4.info)

SampleMetaNamesTable = data.frame(
   row.names = names(NeuroGenesis4),
   Type = c('Yaxis','Yaxis','2Dscatter','2Dscatter'),
   XaxisColumn = c("X","DAYx","tSNE_1","tsne1:ch1"),
   YaxisColumn = c("PJDscores","PJDscores","tSNE_2","tsne2:ch1"),
   COLaxisColumn = c("color","colorBYlabelsX","PJDscores","PJDscores"),
   PCHColumn = c("","","","")
)

grp = list(
Shared.All.4 = c(1 : 4),
Shared.bulk.2 = c(1, 2),
Shared.sc.2 = c(3, 4),
Hs.Meisnr.1 = c(1),
Hs.AZ.1 = c(2),
Gesch.1 = c(3),
Telley.1 = c(4)
)
dims = c(2, 2, 2, 2, 2, 2, 2)

lbb = "NeuroGenesis4.p2"

twoStageLCA.out = twoStageLCA(dataset = NeuroGenesis4.afterWrap, group = grp, comp_num = dims)

SJDScorePlotter.obj = SJDScorePlotter(
    SJDalg = "twoStageLCA",
    scores = twoStageLCA.out$score_list,
    lbb = lbb,
    info = NeuroGenesis4.info,
    SampleMetaNamesTable = SampleMetaNamesTable
)

assemble.byComponent.obj = assemble.byComponent(
SJDScorePlotter.obj = SJDScorePlotter.obj,
component = c(1, 2),
SJD_algorithm = "twoStageLCA",
group = 'Shared.All.4')

}
\keyword{component,}
\keyword{images}
