% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.byDataset.R
\name{assemble.byDataset}
\alias{assemble.byDataset}
\title{Assemble files based on Dataset}
\usage{
assemble.byDataset(
  SJDScorePlotter.obj,
  dataset_name,
  SJD_algorithm,
  group = NA
)
}
\arguments{
\item{SJDScorePlotter.obj}{A list outputted by the SJDScorePlotter function}

\item{dataset_name}{dataset/study analyzed by SJD}

\item{SJD_algorithm}{SJD_algorithm name of SJD algorithm, i.e. concatICA}

\item{group}{group name of the weights group from the SJD_algorithm output, i.e 'Shared.All.13'}
}
\value{
a list of images filtered by dataset
}
\description{
Assemble individual figures for a single dataset analyzed and plotted by SJD
}
\examples{

library(ggplot2)

data(NeuroGenesis4.afterWrap)
data(NeuroGenesis4.info)

SampleMetaNamesTable = data.frame(
   row.names = names(NeuroGenesis4.afterWrap),
   Type = c('Yaxis','Yaxis','2Dscatter','2Dscatter'),
   XaxisColumn = c("X","DAYx","tSNE_1","tsne1:ch1"),
   YaxisColumn = c("PJDscores","PJDscores","tSNE_2","tsne2:ch1"),
   COLaxisColumn = c("color","colorBYlabelsX","PJDscores","PJDscores"),
   PCHColumn = c("","","","")
)

grp = list(
Shared.All.4 = c(1 : 4),
Shared.bulk.2 = c(1, 2),
Shared.sc.2 = c(3, 4),
Hs.Meisnr.1 = c(1),
Hs.AZ.1 = c(2),
Gesch.1 = c(3),
Telley.1 = c(4)
)
dims = c(2, 2, 2, 2, 2, 2, 2)

lbb = "NeuroGenesis4.p2"

twoStageLCA.out = twoStageLCA(dataset = NeuroGenesis4.afterWrap, group = grp, comp_num = dims)

SJDScorePlotter.obj = SJDScorePlotter(
    SJDalg = "twoStageLCA",
    scores = twoStageLCA.out$score_list,
    lbb = lbb,
    info = NeuroGenesis4.info,
    SampleMetaNamesTable = SampleMetaNamesTable
)

assemble.byDataset.obj = assemble.byDataset(
SJDScorePlotter.obj = SJDScorePlotter.obj,
dataset_name = "Meissner.inVitro.bulk.Hs",
SJD_algorithm = "twoStageLCA",
group = NA)

}
\keyword{dataset,}
\keyword{images}
