% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceData.R
\name{balanceData}
\alias{balanceData}
\title{Balance Data}
\usage{
balanceData(dataset)
}
\arguments{
\item{dataset}{A list of data sets to be used}
}
\value{
A list of rebalanced data sets
}
\description{
Balance Data for Concatenate and Joint Analysis
}
\examples{
dataset = list(matrix(c(1 : 8), nrow = 2), matrix(1 : 6, nrow = 2))
balanceData(dataset)

}
\keyword{balance,}
\keyword{dataset}
