% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEMA.R
\name{BEMA}
\alias{BEMA}
\title{BEMA for the standard spiked covariance model}
\usage{
BEMA(eigenvalue, p, n, alpha = 0.2, beta = 0.1)
}
\arguments{
\item{eigenvalue}{a list of eigenvalues to choose from}

\item{p}{dimension of the features}

\item{n}{number of samples}

\item{alpha}{a tuning parameter in the analysis, a default value is set to 0.2}

\item{beta}{a tuning parameter on computing quantile in Tracy-Widom, a default value is set to be 0.1}
}
\value{
The total number of spikes extracted, K
}
\description{
Apply BEMA algorithm for spiked covariance proposed in the paper "Estimation of the number of spiked eigenvalues in a covariance matrix by bulk eigenvalue matching analysis""
}
\examples{
x = matrix(rnorm(1000, 100), nrow = 1000)
eigen_x = svd(x)
eigen_out = list(eigenvalue = eigen_x$d^2 / 100, p = 1000, n = 100)
BEMA(eigen_out$eigenvalue, p = 1000, n = 100)

}
\keyword{Spike}
\keyword{numbers}
