% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAssignment.R
\name{compNameAssign}
\alias{compNameAssign}
\title{Components Name Assignment for concatenate, joint and twoStageLCA analysis}
\usage{
compNameAssign(linked_component_list, group_name)
}
\arguments{
\item{linked_component_list}{list of components extracted}

\item{group_name}{A vector of names for the datasets}
}
\value{
renamed list of linked_component_list
}
\description{
Assign name to components in the concatenate, joint and twoStageLCA methods
}
\examples{
linked_component_list = list(matrix(c(1:4), nrow = 2), matrix(c(1:4), nrow = 2))
group_name = c("x", "y")
compNameAssign(linked_component_list, group_name)

}
\keyword{component,}
\keyword{name}
