% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAssignment.R
\name{compNameAssignSep}
\alias{compNameAssignSep}
\title{Components Name Assignment for Seperate Analysis}
\usage{
compNameAssignSep(linked_component_list, dataset_name)
}
\arguments{
\item{linked_component_list}{list of components extracted}

\item{dataset_name}{A vector of names for the datasets}
}
\value{
renamed list of linked_component_list
}
\description{
Assign name to components in the seperate analysis, sepPCA, sepICA, sepNMF
}
\examples{
linked_component_list = list(matrix(c(1:4), nrow = 2), matrix(c(1:4), nrow = 2))
dataset_name = c("x", "y")
compNameAssignSep(linked_component_list, dataset_name)

}
\keyword{component,}
\keyword{name}
