% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatICA.R
\name{concatICA}
\alias{concatICA}
\title{Concatenated decomposition with Independent Component Analysis}
\usage{
concatICA(
  dataset,
  group,
  comp_num,
  weighting = NULL,
  proj_dataset = NULL,
  proj_group = NULL
)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{weighting}{Weighting of each dataset, initialized to be NULL}

\item{proj_dataset}{The datasets to be projected on}

\item{proj_group}{The grouping of projected data sets}
}
\value{
A list contains the component and the score of each dataset on every component after concatPCA algorithm
}
\description{
Concatenated decomposition of several linked matrices with Independent Component Analysis (ICA)
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1,2,3,4), c(1,2), c(3,4), c(1,3), c(2,4), c(1), c(2), c(3), c(4))
comp_num = c(2,2,2,2,2,2,2,2,2)
proj_dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
proj_group = list(c(TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE))
res_concatICA = concatICA(
dataset,
group,
comp_num,
proj_dataset = proj_dataset,
proj_group = proj_group)

}
\keyword{ICA}
\keyword{pairwise,}
