% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatNMF.R
\name{concatNMF}
\alias{concatNMF}
\title{Concatenated decomposition with Nonnegative Matrix Factorization}
\usage{
concatNMF(
  dataset,
  group,
  comp_num,
  weighting = NULL,
  perturbation = 1e-04,
  proj_dataset = NULL,
  proj_group = NULL,
  enable_normalization = TRUE,
  column_sum_normalization = FALSE,
  screen_prob = NULL
)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{weighting}{Weighting of each dataset, initialized to be NULL}

\item{perturbation}{the perturbation of the 0 element in the analysis}

\item{proj_dataset}{The dataset(s) to be projected on.}

\item{proj_group}{A listed of boolean combinations indicating which groupings should be used for each projected dataset.The length of proj_group should match the length of proj_dataset, and the length of each concatenated boolean combination should match the length of the parameter group.}

\item{enable_normalization}{An argument to decide whether to use normalizaiton or not,  default is TRUE}

\item{column_sum_normalization}{An argument to decide whether to use column sum normalization or not, default it FALSE}

\item{screen_prob}{A vector of probabilies for genes to be chosen}
}
\value{
A list contains the component and the score of each dataset on every component after concatNMF algorithm
}
\description{
Concatenated decomposition of several matrices with Nonnegative Matrix Factorization (NMF)
}
\examples{
Example 1 (1 matrix in proj_dataset):
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1,2,3,4), c(1,2), c(3,4), c(1,3), c(2,4), c(1), c(2), c(3), c(4))
comp_num = c(2,2,2,2,2,2,2,2,2)
proj_dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
proj_group = list(c(TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE))
res_concatNMF = concatNMF(
dataset,
group,
comp_num,
proj_dataset = proj_dataset,
proj_group = proj_group)


Example 2 (multiple matrices in proj_dataset):
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1,2,3,4), c(1,2), c(3,4), c(1,3), c(2,4), c(1), c(2), c(3), c(4))
comp_num = c(2,2,2,2,2,2,2,2,2)
proj_dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))   # length(proj_dataset should match length(proj_group))
proj_group = list(c(TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE), c(TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE))
res_concatNMF = concatNMF(
dataset,
group,
comp_num,
proj_dataset = proj_dataset,
proj_group = proj_group)

}
\keyword{NMF}
\keyword{pairwise,}
