% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunction.R
\name{configuration_setting_generation}
\alias{configuration_setting_generation}
\title{Configuration for simulated data generation}
\usage{
configuration_setting_generation(
  featureNum = 50,
  DataNum = c(100, 100, 100, 100),
  commonlySharedNum = 2,
  partiallySharedNum = c(2, 2, 2, 2),
  individualSharedNum = c(2, 2, 2, 2),
  noiseVariance = c(1, 1, 1, 1)
)
}
\arguments{
\item{featureNum}{number of features}

\item{DataNum}{number of data}

\item{commonlySharedNum}{number of common component}

\item{partiallySharedNum}{number of partial shared component}

\item{individualSharedNum}{number of individual component}

\item{noiseVariance}{variance of noise}
}
\description{
generate the configuration for the data
}
\examples{
configuration_setting_generation()

}
\keyword{configure}
\keyword{data}
\keyword{for}
\keyword{generation}
\keyword{of}
\keyword{setting}
\keyword{the}
