% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameExtractor.R
\name{datasetNameExtractor}
\alias{datasetNameExtractor}
\title{Dataset Name Extractor}
\usage{
datasetNameExtractor(dataset)
}
\arguments{
\item{dataset}{A list of data sets}
}
\value{
A vector of strings of dataset name
}
\description{
Extract data.frame name from a list of data.frames, if
no name is given in the list, it will be renamed as
"dataset_no.index"
}
\examples{
dataset = list(
x = matrix(c(1 : 4), nrow = 2),
y = matrix(c(1 : 4), nrow = 2))

datasetNameExtractor(dataset)

}
\keyword{name}
