% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterNAValue.R
\name{filterNAValue}
\alias{filterNAValue}
\title{Filter NA Value}
\usage{
filterNAValue(list_score, dataset, group)
}
\arguments{
\item{list_score}{A list of scores extracted in the analysis}

\item{dataset}{A list of datasets used in the analysis}

\item{group}{A list of group assignments}
}
\value{
A list of scores by assigning NA to scores not in groups
}
\description{
Assign NA values to scores not in the group for each data set
}
\examples{

x = list(matrix(c(1 : 4), nrow = 2), matrix(c(1 : 4), nrow = 2))
y = list(matrix(c(1 : 4), nrow = 2), matrix(c(1 : 4), nrow = 2))
list_score = list(x, y)
dataset = c(1, 2)
group = list(c(1), c(2))
filterNAValue(list_score, dataset, group)

}
\keyword{NA,}
\keyword{filter}
