% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexNMF.R
\name{flexNMF}
\alias{flexNMF}
\title{Flexible Decomposition with Nonnegative Matrix Factorization}
\usage{
flexNMF(
  dataset,
  gene_score = NULL,
  sample_score = NULL,
  normalize = TRUE,
  max_ite = 1000,
  max_err = 1e-04,
  enable_normalization = TRUE,
  column_sum_normalization = FALSE
)
}
\arguments{
\item{dataset}{The dataset to be decomposed}

\item{gene_score}{The known gene embedding columns/patterns to be fixed}

\item{sample_score}{The known sample score columns/patterns to be fixed}

\item{normalize}{whether the dataset matrix needs to be normalized}

\item{max_ite}{The maximum number of iterations for the jointNMF algorithms to run, default value is set to 1000}

\item{max_err}{The maximum error of loss between two iterations, or the program will terminate and return, default value is set to be 0.0001}

\item{enable_normalization}{An argument to decide whether to use normalizaiton or not,  default is TRUE}

\item{column_sum_normalization}{An argument to decide whether to use column sum normalization or not, default it FALSE}
}
\value{
A list that contains the 1] gene embedding scores of the decomposed dataset. 2] sample scores of the decmposed dataset. and 3] the log of errors as the NMF was iterated.
}
\description{
Function to decompose a matrix with known gene loading and/or sample embedding with Nonnegative Matrix Factorization.
}
\details{
flexNMF decomposes the matrix while keeps the given/known gene columns or sample rows fixed, estimating the rest of the gene and sample matrices. 


dataset, gene_score=NULL, sample_score=NULL, normalize = TRUE, max_ite = 1000, max_err = 0.0001, enable_normalization = TRUE, column_sum_normalization = FALSE
}
\examples{
dataset = matrix(runif(5000, 1, 2), nrow = 100, ncol = 50)
gene_score = matrix(runif(200, 1, 2), nrow = 100, ncol = 2)
sample_score = matrix(runif(100, 1, 2), nrow = 2, ncol = 50)
normalize = TRUE
res_flexNMF = flexNMF(
dataset = dataset,
gene_score = gene_score,
sample_score = sample_score,
normalize = normalize)


}
\keyword{NMF,}
\keyword{decomposition}
\keyword{flexible,}
