% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameToMatrix.R
\name{frameToMatrix}
\alias{frameToMatrix}
\title{Transform data.frame to matrix}
\usage{
frameToMatrix(dataset)
}
\arguments{
\item{dataset}{A list of data sets}
}
\value{
A list of matrix transformed from the list of data frames
}
\description{
Function to transform a data frame into a matrix
}
\examples{
dataset = list(
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50)
)
frameToMatrix(dataset)

}
\keyword{dataframe,}
\keyword{matrix}
