% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAssignment.R
\name{geneNameAssign}
\alias{geneNameAssign}
\title{Gene name assignment}
\usage{
geneNameAssign(linked_component_list, gene_name)
}
\arguments{
\item{linked_component_list}{A list of extracted components}

\item{gene_name}{A vector of strings for gene names}
}
\value{
A list of components with gene name added
}
\description{
Assign gene names to component derived in the analysis
}
\examples{
x = matrix(c(1 : 4), nrow = 2)
y = matrix(c(1 : 4), nrow = 2)
component_list = list(x, y)
gene_name = c("gene.1", "gene.2")
geneNameAssign(component_list, gene_name)

}
\keyword{gene,}
\keyword{name}
