% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameExtractor.R
\name{geneNameExtractor}
\alias{geneNameExtractor}
\title{Gene Name Extractor}
\usage{
geneNameExtractor(dataset)
}
\arguments{
\item{dataset}{A list of data sets to be analyzed}
}
\value{
a vector of strings of gene name
}
\description{
Extract gene names from the input dataset
}
\examples{
x = matrix(c(1 : 4), nrow = 2)
rownames(x) = c("row1", "row2")
y = matrix(c(1 : 4), nrow = 2)
rownames(y) = c("row1", "row2")
dataset = list(x, y)
geneNameExtractor(dataset)

}
\keyword{gene,}
\keyword{name}
