% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScreen.R
\name{geneScreen}
\alias{geneScreen}
\title{Gene Screen}
\usage{
geneScreen(dataset, screen_prob)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{screen_prob}{A vector of probabilies for genes to be chosen}
}
\value{
A list contains the component and the score of each dataset on every component after jointPCA algorithm
}
\description{
Screen genes based on their standard deviation
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))

screen_prob = c(0.2, 0.2, 0.2, 0.2)

screened_dataset = geneScreen(dataset, screen_prob)

}
\keyword{gene}
\keyword{screen,}
