% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameExtractor.R
\name{groupNameExtractor}
\alias{groupNameExtractor}
\title{Group Name Extractor}
\usage{
groupNameExtractor(group)
}
\arguments{
\item{group}{A list of group assignments for the data sets with group name on it}
}
\value{
A vector of strings of group name
}
\description{
Extract group name from groups, if no name is given in the input
list, the output vector will be represented as
"component_No.index"
}
\examples{
dataset = list(
x = c(1, 2, 3),
y = c(1, 2, 4))

}
\keyword{name}
