% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointFlexNMF.R
\name{jointFlexNMF}
\alias{jointFlexNMF}
\title{Joint Flexible Decomposition with Nonnegative Matrix Factorization}
\usage{
jointFlexNMF(
  dataset,
  group,
  comp_num,
  gene_score = NULL,
  sample_score = NULL,
  weighting = NULL,
  max_ite = 1000,
  max_err = 1e-04,
  enable_normalization = TRUE,
  column_sum_normalization = FALSE,
  screen_prob = NULL
)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{gene_score}{The known gene embedding columns/patterns to be fixed}

\item{sample_score}{The known sample score columns/patterns to be fixed}

\item{weighting}{Weighting of each dataset, initialized to be NULL}

\item{max_ite}{The maximum number of iterations for the jointNMF algorithms to run, default value is set to 100}

\item{max_err}{The maximum error of loss between two iterations, or the program will terminate and return, default value is set to be 0.0001}

\item{enable_normalization}{An argument to decide whether to use normalizaiton or not,  default is TRUE}

\item{column_sum_normalization}{An argument to decide whether to use column sum normalization or not, default it FALSE}

\item{screen_prob}{A vector of probabilies for genes to be chosen}
}
\value{
A list contains the component and the score of each dataset on every component after jointFlexNMF algorithm
}
\description{
Joint flexible decomposition of several linked matrices with Nonnegative Matrix Factorization (NMF)
It combines the jointNMF function and the flexNMF function, allowing flexible decomposition of several matrices with known patterns (gene embeddings and/or sample scores).
It is based on the MSE loss, proposed by Lee, Daniel D., and H. Sebastian Seung. "Learning the parts of objects by non-negative matrix factorization." Nature 401.6755 (1999): 788-791.
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1,2,3,4), c(1,2), c(3,4), c(1,3), c(2,4), c(1), c(2), c(3), c(4))
comp_num = c(2,2,2,2,2,2,2,2,2)
gene_score = matrix(runif(200, 1, 2), nrow = 100, ncol = 2)
sample_score = matrix(runif(400, 1, 2), nrow = 2, ncol = 200)
res_jointFlexNMF = jointFlexNMF(
dataset,
group,
comp_num,
gene_score = gene_score,
sample_score = sample_score)


}
\keyword{NMF,}
\keyword{decomposition}
\keyword{flexible,}
\keyword{joint,}
