% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointICA.R
\name{jointICA}
\alias{jointICA}
\title{Joint decomposition with Independent Component Analysis}
\usage{
jointICA(
  dataset,
  group,
  comp_num,
  weighting = NULL,
  max_ite = 100,
  max_err = 1e-04,
  proj_dataset = NULL,
  proj_group = NULL,
  enable_normalization = TRUE,
  column_sum_normalization = FALSE,
  screen_prob = NULL
)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{weighting}{Weighting of each dataset, initialized to be NULL}

\item{max_ite}{The maximum number of iterations for the jointPCA algorithms to run, default value is set to 100}

\item{max_err}{The maximum error of loss between two iterations, or the program will terminate and return, default value is set to be 0.0001}

\item{proj_dataset}{The datasets to be projected on}

\item{proj_group}{The grouping of projected data sets}

\item{enable_normalization}{An argument to decide whether to use normalizaiton or not,  default is TRUE}

\item{column_sum_normalization}{An argument to decide whether to use column sum normalization or not, default it FALSE}

\item{screen_prob}{A vector of probabilies for genes to be chosen}
}
\value{
A list contains the component and the score of each dataset on every component after jointPCA algorithm
}
\description{
Joint decomposition of several linked matrices with Independent Component Analysis (ICA)
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1,2,3,4), c(1,2), c(3,4), c(1,3), c(2,4), c(1), c(2), c(3), c(4))
proj_dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
proj_group = list(c(TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE))
comp_num = c(2,2,2,2,2,2,2,2,2)
res_jointICA = jointICA(
dataset,
group,
comp_num,
proj_dataset = proj_dataset,
proj_group = proj_group)

}
\keyword{ICA}
\keyword{Joint,}
