% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPCA.R
\name{jointPCA}
\alias{jointPCA}
\title{Joint Decomposition with Principal Component Analysis}
\usage{
jointPCA(
  dataset,
  group,
  comp_num,
  weighting = NULL,
  max_ite = 100,
  max_err = 1e-04,
  proj_dataset = NULL,
  proj_group = NULL
)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{weighting}{Weighting of each dataset, initialized to be NULL}

\item{max_ite}{The maximum number of iterations for the jointPCA algorithms to run, default value is set to 100}

\item{max_err}{The maximum error of loss between two iterations, or the program will terminate and return, default value is set to be 0.001}

\item{proj_dataset}{The datasets to be projected on}

\item{proj_group}{The grouping of projected data sets}
}
\value{
A list contains the component and the score of each dataset on every component after jointPCA algorithm
}
\description{
Joint decomposition of several linked matrices with Principal Component Analysis (PCA)
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1,2,3,4), c(1,2), c(3,4), c(1,3), c(2,4), c(1), c(2), c(3), c(4))
comp_num = c(2,2,2,2,2,2,2,2,2)
proj_dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
proj_group = list(c(TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE))
res_jointPCA = jointPCA(dataset, group, comp_num, proj_dataset = proj_dataset, proj_group = proj_group)

}
\keyword{PCA}
\keyword{joint,}
