% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedPCA.R
\name{linkedPCA}
\alias{linkedPCA}
\title{Linked Component Analysis}
\usage{
linkedPCA(dataset, cov_list, eigen_space, group, comp_num)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{cov_list}{A list of covariance of the datasets}

\item{eigen_space}{A matrix of the space of the signal}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}
}
\value{
A list of component
}
\description{
Functions to implement the linked component analysis
}
\keyword{Analysis,}
\keyword{Component}
\keyword{Linked}
\keyword{sequential}
