% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileMarchenkoPastur.R
\name{marchenko_pastur_quantile}
\alias{marchenko_pastur_quantile}
\title{Quantile for Marchenko Pastur Distribution}
\usage{
marchenko_pastur_quantile(q, n, p, step_size = 1e-04)
}
\arguments{
\item{q}{quantile to be chosen}

\item{n}{number of samples to be considered}

\item{p}{dimension of the covariance matrix}

\item{step_size}{step length to do numerical integration}
}
\value{
Quantile location
}
\description{
Compute the q quantile for Marchenko Pastur Distribution with n and p
}
\examples{
out = marchenko_pastur_quantile(0.2, 50, 500)

}
\keyword{Marchenko-Pastur,}
\keyword{quantile}
