% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeData.R
\name{min_max_normalization}
\alias{min_max_normalization}
\title{Min-Max Normalization Function}
\usage{
min_max_normalization(x)
}
\arguments{
\item{x}{A numeric vector to be normalized.}
}
\value{
A normalized version of the input vector with values scaled between 0 and 1.
}
\description{
This function performs min-max normalization on a vector.
}
\examples{

list_score = matrix(runif(5000, 1, 2), nrow = 100, ncol = 50)

}
