% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtx.R
\docType{data}
\name{mtx}
\alias{mtx}
\title{mtx sequence data}
\format{
A list containing two dataframes
}
\usage{
data(mtx)
}
\description{
Data from sequecing, it contains two datasets: HS.Nico and Nm.Nico
HS.Nico is from human and Mm.Nico is from mouse
}
\examples{

data(mtx)
head(mtx$HS.Nico)
head(mtx$Mm.Nico)


}
\keyword{datasets}
