% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroGenesis4.R
\docType{data}
\name{NeuroGenesis4.afterWrap}
\alias{NeuroGenesis4.afterWrap}
\title{NeuroGenesis4.afterWrap}
\format{
A list containing 4 dataframes
}
\usage{
data(NeuroGenesis4.afterWrap)
}
\description{
A list of 4 neuro data sets sampled after wrapper, containing: Meissner.inVitro.bulk.Hs, LIBD.AZ.inVitro.bulk.Hs, Geschwind.inVivo.sc.Hs, Jabaudon.inVivo.sc.Mm
}
\examples{

data(NeuroGenesis4.afterWrap)
head(NeuroGenesis4.afterWrap$Meissner.inVitro.bulk.Hs)
head(NeuroGenesis4.afterWrap$LIBD.AZ.inVitro.bulk.Hs)
head(NeuroGenesis4.afterWrap$Geschwind.inVivo.sc.Hs)
head(NeuroGenesis4.afterWrap$Jabaudon.inVivo.sc.Mm)

}
\keyword{datasets}
