% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroGenesis4.R
\docType{data}
\name{NeuroGenesis4.info}
\alias{NeuroGenesis4.info}
\title{NeuroGenesis4.info}
\format{
A list containing 4 dataframes
}
\usage{
data(NeuroGenesis4)
}
\description{
A list of info for 4 neuro data sets sampled, containing: Meissner.inVitro.bulk.Hs, LIBD.AZ.inVitro.bulk.Hs, Geschwind.inVivo.sc.Hs, Jabaudon.inVivo.sc.Mm
}
\examples{

data(NeuroGenesis4.info)
head(NeuroGenesis4.info$Meissner.inVitro.bulk.Hs)
head(NeuroGenesis4.info$LIBD.AZ.inVitro.bulk.Hs)
head(NeuroGenesis4.info$Geschwind.inVivo.sc.Hs)
head(NeuroGenesis4.info$Jabaudon.inVivo.sc.Mm)

}
\keyword{datasets}
