% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroGenesis4.R
\docType{data}
\name{NeuroGenesis4}
\alias{NeuroGenesis4}
\title{NeuroGenesis4}
\format{
A list containing 4 dataframes
}
\usage{
data(NeuroGenesis4)
}
\description{
A list of 4 neuro data sets sampled, containing: Meissner.inVitro.bulk.Hs, LIBD.AZ.inVitro.bulk.Hs, Geschwind.inVivo.sc.Hs, Jabaudon.inVivo.sc.Mm
}
\examples{

data(NeuroGenesis4)
head(NeuroGenesis4$Meissner.inVitro.bulk.Hs)
head(NeuroGenesis4$LIBD.AZ.inVitro.bulk.Hs)
head(NeuroGenesis4$Geschwind.inVivo.sc.Hs)
head(NeuroGenesis4$Jabaudon.inVivo.sc.Mm)

}
\keyword{datasets}
