% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeData.R
\name{normalizeData}
\alias{normalizeData}
\title{Data Normalization}
\usage{
normalizeData(
  dataset,
  enable_normalization = TRUE,
  column_sum_normalization = TRUE,
  nonnegative_normalization = FALSE
)
}
\arguments{
\item{dataset}{The input list of data sets matrix}

\item{enable_normalization}{An argument to decide whether to use normalizaiton or not,  default is TRUE}

\item{column_sum_normalization}{An argument to decide whether to use column sum normalization or not, default it TRUE}

\item{nonnegative_normalization}{An argument to decide wehther it is nonnegative matrix factorization based or not, default is FALSE}
}
\description{
Normalize data to have mean zero and std 1
}
\examples{
dataset = list(
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50)
)
normalizeData(dataset)

}
\keyword{normalize}
