% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjdWrapper.R
\name{pjd_wrap}
\alias{pjd_wrap}
\title{Wrapper function for PJD package}
\usage{
pjd_wrap(
  data.list,
  species.vector,
  geneType.vector,
  geneType.out = "symbol",
  species.out
)
}
\arguments{
\item{data.list}{Input list of data sets}

\item{species.vector}{A vector of species types}

\item{geneType.vector}{A vector of gene types,}

\item{geneType.out}{Type of genes}

\item{species.out}{Kind of species output}
}
\value{
A list of data sets, after alignment
}
\description{
A wrapper function to align multiple datasets from different species
}
\examples{

data(mtx)

PJD_example = pjd_wrap(data.list = mtx,
species.vector = c("human", "mouse"),
geneType.vector = c("symbol", "symbol"),
geneType.out = "symbol",
species.out = "human")

}
\keyword{PJD}
\keyword{Wrapper,}
