% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPCA.R
\name{Procrustes}
\alias{Procrustes}
\title{Procrustes Projection}
\usage{
Procrustes(A, B)
}
\arguments{
\item{A}{The input matrix A as target}

\item{B}{the input matrix B as basis}
}
\value{
The procrustes matrix U
}
\description{
Procrustes projection function to solve the procrustes problem ||A - UB||^2 U^T U = I
}
\keyword{procrustes}
