% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectiLCA.R
\name{projectiLCA}
\alias{projectiLCA}
\title{Function to estimate sample embeddings for one dataset from a gene loading matrix derived from an iLCA analysis of another dataset.}
\usage{
projectiLCA(
  proj_dataset,
  proj_group,
  list_component,
  ica_score,
  max_ite = 1000,
  max_err = 1e-04,
  enable_normalization = TRUE,
  column_sum_normalization = FALSE
)
}
\arguments{
\item{proj_dataset}{The dataset(s) to be projected on.}

\item{proj_group}{A logical vector indicating which groupings, i. e. which elements of list_component should be used for each projected dataset. The length of proj_group should match the length of list_component.}

\item{list_component}{a single matrix of gene loadings as a list element, or a list_component produced from a twoStageiLCA() decomposition.}

\item{ica_score}{ice_score produced from a twoStageiLCA() decomposition.}

\item{max_ite}{The maximum number of iterations for the twoStageiLCA algorithms to run, default value is set to 1000}

\item{max_err}{The maximum error of loss between two iterations, or the program will terminate and return, default value is set to be 0.0001}

\item{enable_normalization}{An argument to decide whether to use normalizaiton or not,  default is TRUE}

\item{column_sum_normalization}{An argument to decide whether to use column sum normalization or not, default it FALSE}
}
\value{
A list that contains the projected scores of each dataset on every component.
}
\description{
projectiLCA estimates the embeddings for samples in a new dataset when given a gene loading matrix from an iLCA analysis result of another single matrix, 
or set of matrices (e.g. the "list_component" from a twoStageiLCA output object)
}
\details{
PLEASE MAKE SURE YOUR proj_dataset AND list_component ELEMENTS HAVE MEANINGFUL ROW(GENE) NAMES - they are matched across matrices for the projection.
}
\examples{
proj_dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
proj_group = c(TRUE, TRUE) # which groupings in the twoStageiLCA analysis you want to project on.
list_component = twoStageiLCA_res$linked_component_list # from twoStageiLCA result
ica_score = twoStageiLCA_res$ica_score # from twoStageiLCA result
res_projiLCA = projectiLCA(
proj_dataset = proj_dataset,
proj_group = proj_group,
list_component = list_component,
ica_score = ica_score)



}
\keyword{iLCA}
\keyword{projection,}
\keyword{twoStageiLCA,}
