% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pve.R
\name{pveMultiple}
\alias{pveMultiple}
\title{Percentage of Variance Explained for Multiple Data sets}
\usage{
pveMultiple(dataset, group, comp_num, list_score, list_component)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{list_score}{A list of extracted scores by the corresponding algorithm}

\item{list_component}{A list of components comptuted by the corresponding algorithm}
}
\value{
The list of scores
}
\description{
Compute the PVE (Percentage of Variance Explained) for multiple data sets on multiple components
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1,2,3,4), c(1,2), c(3,4), c(1,3), c(2,4), c(1), c(2), c(3), c(4))
comp_num = c(2,2,2,2,2,2,2,2,2)
res_concatPCA = concatPCA(dataset, group, comp_num)
pveMultiple(dataset, group, comp_num, res_concatPCA$score_list, res_concatPCA$linked_component_list)

}
\keyword{Multiple}
\keyword{PVE,}
\keyword{analysis}
