% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pve.R
\name{pveSep}
\alias{pveSep}
\title{Percentage of Variance Explained for separate data set}
\usage{
pveSep(dataset, list_score, list_component)
}
\arguments{
\item{dataset}{A list of data sets for input}

\item{list_score}{A list of extracted scores by the corresponding algorithm}

\item{list_component}{A list of components comptuted by the corresponding algorithm}
}
\value{
The list of scores
}
\description{
Compute the PVE (percentage of variance explained) for each data set
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepPCA = sepPCA(dataset, comp_num)
pveSep(dataset, res_sepPCA$score_list, res_sepPCA$linked_component_list)

}
\keyword{PVE,}
\keyword{Separate}
\keyword{analysis}
