% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceData.R
\name{rebalanceData}
\alias{rebalanceData}
\title{Rebalance Data}
\usage{
rebalanceData(list_score, group, dataset)
}
\arguments{
\item{list_score}{A list of scores extracted in the analysis}

\item{group}{A list of group assignment}

\item{dataset}{A list of dataset in analysis}
}
\value{
A list of rebalanced scores
}
\description{
Rebalance scores based on the balanced data set
}
\examples{
x = list(matrix(c(1 : 4), nrow = 2), matrix(c(1 : 4), nrow = 2))
y = list(matrix(c(1 : 4), nrow = 2), matrix(c(1 : 4), nrow = 2))
list_score = list(x, y)
group = list(c(1), c(2))
dataset = list(matrix(c(1 : 8), nrow = 2), matrix(1 : 6, nrow = 2))
rebalanceData(list_score, group, dataset)

}
