% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunction.R
\name{rotate_component}
\alias{rotate_component}
\title{Simulated Component Rotation}
\usage{
rotate_component(comp, angle = 0)
}
\arguments{
\item{comp}{A matrix of components}

\item{angle}{Rotation angle of the component}
}
\value{
Matrix of rotated component
}
\description{
Rotated the simulated component generated
}
\examples{
component = svd(matrix(rnorm(100 * 200), nrow = 200))$u[, 1 : 2]
rotate_component(component, pi / 6)

}
\keyword{component}
\keyword{rotation,}
