% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAssignment.R
\name{sampleNameAssign}
\alias{sampleNameAssign}
\title{Sample Name Assignment for Concatenate, Joint and TwoStageLCA Analysis}
\usage{
sampleNameAssign(score_list, sample_name)
}
\arguments{
\item{score_list}{A list of scores to do analysis}

\item{sample_name}{A list of names for samples to be analyzed}
}
\value{
A list of scores to analyze
}
\description{
Assign sample names to the scores
}
\examples{
x = list(matrix(c(1 : 4), nrow = 2), matrix(c(1 : 4), nrow = 2))
y = list(matrix(c(1 : 4), nrow = 2), matrix(c(1 : 4), nrow = 2))
score_list = list(x, y)
sample_name = list(c("x.sample.1", "x.sample.2"), c("y.sample.1", "y.sample.2"))
sampleNameAssign(score_list, sample_name)

}
\keyword{name}
\keyword{sample,}
