% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAssignment.R
\name{sampleNameAssignSep}
\alias{sampleNameAssignSep}
\title{Sample Name Assignment for Seperate Analysis}
\usage{
sampleNameAssignSep(score_list, sample_name)
}
\arguments{
\item{score_list}{List of score for each data.frame}

\item{sample_name}{List of names for the samples in the list}
}
\value{
A list of scores for the samples
}
\description{
Assign sample names to each dataset in the list of data sets,
it takes two arguments, the first is the list of scores,
the second is the list of sample names
}
\examples{
x = matrix(c(1:4), nrow = 2)
y = matrix(c(1:4), nrow = 2)
score_list = list(x, y)
sample_name = list(c("x.sample.1", "x.sample.2"), c("y.sample.1", "y.sample.2"))
sampleNameAssignSep(score_list, sample_name)

}
\keyword{name}
\keyword{sample,}
