% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameExtractor.R
\name{sampleNameExtractor}
\alias{sampleNameExtractor}
\title{Sample name extractor}
\usage{
sampleNameExtractor(dataset)
}
\arguments{
\item{dataset}{A list of datasets containing sample names}
}
\value{
A vector of sample names
}
\description{
Extract a list of sample names from input list of datasets
}
\examples{
x = matrix(c(1 : 4), nrow = 2)
colnames(x) = c("sp1", "sp2")
y = matrix(c(1 : 4), nrow = 2)
colnames(y) = c("sp3", "sp4")
dataset = list(x, y)
sampleNameExtractor(dataset)

}
\keyword{extractor}
\keyword{name,}
