% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunction.R
\name{score_generation}
\alias{score_generation}
\title{Random Score Generation}
\usage{
score_generation(dim_score, num_score, score_variance)
}
\arguments{
\item{dim_score}{dimension of the scores}

\item{num_score}{number of scores}

\item{score_variance}{variance of the score}
}
\value{
a matrix of generated random scores, with dim_score * num_score
}
\description{
Gererate random scores for the simulation
}
\examples{
score_generation(2, 10, c(1,2))

}
\keyword{random}
\keyword{score}
