% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAssignment.R
\name{scoreNameAssign}
\alias{scoreNameAssign}
\title{Score Name Assignment for Concatenate, Joint and TwoStageLCA Analysis}
\usage{
scoreNameAssign(score_list, dataset_name, group_name)
}
\arguments{
\item{score_list}{A list of scores in the analysis}

\item{dataset_name}{A vector of datasets names extracted by dataNameExtractor function}

\item{group_name}{A vector of group names extracted by groupNameExtractor function}
}
\value{
A list of scores assigned with names
}
\description{
Assign names to score lists based on dataset_name and group_name,
the dataset_name is the output of the dataNameExtractor, the group_name
is the output of the groupNameExtractor
}
\examples{
score_list = list(
list(matrix(c(1 : 4), nrow = 2), matrix(c(1 : 4), nrow = 2)),
list(matrix(c(1 : 4), nrow = 2), matrix(c(1 : 4), nrow = 2)))
dataset_name = c("dat1", "dat2")
group_name = c("comp1", "comp2")
scoreNameAssign(score_list, dataset_name, group_name)

}
\keyword{name}
\keyword{score,}
