% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAssignment.R
\name{scoreNameAssignSep}
\alias{scoreNameAssignSep}
\title{Score Name Assign for Seperate Analysis}
\usage{
scoreNameAssignSep(score_list, dataset_name)
}
\arguments{
\item{score_list}{List of scores in sep analysis}

\item{dataset_name}{List of dataset names extracted by datasetNameExtractor}
}
\value{
A list of scores, content same as input name changed based on dataset_name
}
\description{
Assign name to scores based on the extracted name for each dataset,
it takes the vector of dataset_name outputed by  the 'datasetNameExtractor'
function
}
\examples{
score_list = list(matrix(c(1 : 4), nrow = 2), matrix(c(5 : 8), nrow = 2))
dataset_name = c("x", "y")
scoreNameAssignSep(score_list, dataset_name)

}
\keyword{name}
\keyword{score,}
