% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreePlot_LCvsPC.R
\name{ScreePlot_LCvsPC}
\alias{ScreePlot_LCvsPC}
\title{genetrate scree plot using tsLCA and sepPCA embeddings for each matrix}
\usage{
ScreePlot_LCvsPC(LCAobj, group, sepPCAobj)
}
\arguments{
\item{LCAobj}{list output of tsLCA run, list}

\item{group}{name of the weights group from the SJD_algorithm output, i.e 'Shared.All.13', str}

\item{sepPCAobj}{list output of sepPCA run, list}
}
\description{
genetrate scree plot using tsLCA and sepPCA embeddings for each matrix
}
