% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepICA.R
\name{sepICA}
\alias{sepICA}
\title{Single Data Set Decomposition with Independent Component Analysis}
\usage{
sepICA(
  dataset,
  comp_num,
  weighting = NULL,
  enable_normalization = TRUE,
  column_sum_normalization = TRUE,
  screen_prob = NULL
)
}
\arguments{
\item{dataset}{A dataframe/matrix to be decomposed}

\item{comp_num}{Number of ICs to be extracted}

\item{weighting}{Weighting of each dataset, initialized to be NULL}

\item{enable_normalization}{An argument to decide whether to use normalizaiton or not,  default is TRUE}

\item{column_sum_normalization}{An argument to decide whether to use column sum normalization or not, default it TRUE}

\item{screen_prob}{A vector of probabilies for genes to be chosen}
}
\value{
A list of scores and component
}
\description{
Apply ICA (Independent Component Analysis) to a single data set
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepICA = sepICA(dataset, comp_num)

}
\keyword{ICA}
\keyword{analysis,}
\keyword{separate}
