% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunction.R
\name{simulated_data_generation}
\alias{simulated_data_generation}
\title{Simulated Data Generation}
\usage{
simulated_data_generation(
  configuration_setting,
  amplitude = 1,
  heterogeneousNoise = FALSE
)
}
\arguments{
\item{configuration_setting}{setting for the configuration}

\item{amplitude}{The amplitude of the score variance}

\item{heterogeneousNoise}{Whether the noise for each dataset to be heteregeneous}
}
\description{
Generate simulation data
}
\examples{
configuration_setting = configuration_setting_generation()
simulated_data_generation(configuration_setting)

}
\keyword{data}
\keyword{generation}
\keyword{simulated}
