% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjdWrap.R
\name{sjdWrap}
\alias{sjdWrap}
\title{SJD Wrap}
\usage{
sjdWrap(
  data.list,
  species.vector,
  geneType.vector,
  geneType.out = "symbol",
  species.out,
  useNewestVersion = FALSE
)
}
\arguments{
\item{data.list}{input list of expression matrices from different species i.e human and mouse datasets}

\item{species.vector}{character of species type of each matrix i.e c('human', 'mouse', 'mouse')}

\item{geneType.vector}{character of gene/rowname type of each matrix i.e c("symbol","ensembl","symbol")}

\item{geneType.out}{character of output gene/rowname type of each matrix i.e "symbol"}

\item{species.out}{character of output species type for gene/rowname}

\item{useNewestVersion}{logical indicating if function getMatch should attempt to use the latest version of ensembl, or to use the May 2025 archive version that is more stable. default is FALSE.}
}
\value{
A list of expression matrices (of different species) with only shared genes
}
\description{
wrapping up expression matrices (of different species) with only shared genes as SJD input
}
\examples{

data(NeuroGenesis4)
SJDdataIN = sjdWrap(
data.list = NeuroGenesis4,
species.vector=c("human","human","human","mouse"),
geneType.vector=c("symbol","ensembl","symbol","symbol"),
geneType.out="symbol",
species.out="human")

}
\keyword{expression}
\keyword{gene}
\keyword{shared}
