% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjdWrapProjection.R
\name{sjdWrapProjection}
\alias{sjdWrapProjection}
\title{SJD Wrap for projection}
\usage{
sjdWrapProjection(
  data.list.template,
  data.list.projection,
  species.template,
  species.vec.projection,
  geneType.template,
  geneType.vec.projection
)
}
\arguments{
\item{data.list.template}{input list of expression matrices from the output of the sjdWrap function}

\item{data.list.projection}{input list of expression matrices needed to be matched with data.list.template}

\item{species.template}{character of species type from the sjdWrap function}

\item{species.vec.projection}{list of species of each matrix to be projected}

\item{geneType.template}{character of gene/rowname type from the sjdWrap function}

\item{geneType.vec.projection}{list of output gene/rowname type of each matrix to be projected}
}
\value{
A list of expression matrices (of different species) with only shared genes to be projected
}
\description{
Wrapping up a new list of projection matrices with the list computed from the function of sjdWrap
}
\examples{

## Load NeuroGenesis4 data into R
data(NeuroGenesis4)

## sjdWrap of the training data sets
SJDdataIN = sjdWrap(
data.list = NeuroGenesis4,
species.vector=c("human","human","human","mouse"),
geneType.vector=c("symbol","ensembl","symbol","symbol"),
geneType.out="symbol",
species.out="human")

## Sample from data, serving as the projection expression matrices
NeuroGenesis4.sample = NeuroGenesis4
NeuroGenesis4.sample[[1]] = NeuroGenesis4.sample[[1]][-5,]
rownames(NeuroGenesis4.sample[[1]])[5] = paste0(rownames(NeuroGenesis4.sample[[1]])[5], ".test")

NeuroGenesis4.sample[[2]] = NeuroGenesis4.sample[[2]][-10,]
rownames(NeuroGenesis4.sample[[2]])[10] = paste0(rownames(NeuroGenesis4.sample[[2]])[10], ".test")

NeuroGenesis4.sample[[3]] = NeuroGenesis4.sample[[3]][-15,]
rownames(NeuroGenesis4.sample[[3]])[15] = paste0(rownames(NeuroGenesis4.sample[[3]])[15], ".test")

NeuroGenesis4.sample[[4]] = NeuroGenesis4.sample[[4]][-20,]
rownames(NeuroGenesis4.sample[[4]])[20] = paste0(rownames(NeuroGenesis4.sample[[4]])[20], ".test")

SJDdataProjection = sjdWrapProjection(SJDdataIN, NeuroGenesis4.sample, "human", c("human","human","human","mouse"), "symbol", c("symbol","ensembl","symbol","symbol"))

}
\keyword{expression}
\keyword{gene}
\keyword{shared}
