% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSJDtoSeurat.R
\name{store_to_seurat}
\alias{store_to_seurat}
\title{Add SJD loadings to a Seurat object}
\usage{
store_to_seurat(Seurat.obj, SJDoutput, Dataset, SJDloading, SJDmethod = NA)
}
\arguments{
\item{Seurat.obj}{A Seurat Object.}

\item{SJDoutput}{A SJD object of decomposition results.}

\item{Dataset}{The name of the Dataset of interests in the SJD score_list component}

\item{SJDloading}{The name of the SJDloading of interests in the SJD score_list component under Dataset.}

\item{SJDmethod}{The name of the SJD method used for the decomposition, e.g. "twoStageLCA". If NA, will use "SJD"}
}
\description{
Add SJD loadings to a Seurat object
}
