% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageiLCA.rank.R
\name{twoStageiLCA.rank}
\alias{twoStageiLCA.rank}
\title{Two-staged Independent LCA and automatic rank selection}
\usage{
twoStageiLCA.rank(
  dataset,
  group,
  weighting = NULL,
  total_number = NULL,
  threshold,
  backup = 0,
  plotting = FALSE,
  proj_dataset = NULL,
  proj_group = NULL,
  enable_normalization = TRUE,
  column_sum_normalization = TRUE,
  screen_prob = NULL
)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{weighting}{Weighting of each dataset, initialized to be NULL}

\item{total_number}{Total number of components will be extracted, if default value is set to NA, then BEMA will be used.}

\item{threshold}{The threshold used to cutoff the eigenvalues}

\item{backup}{A backup variable, which permits the overselection of the components by BEMA}

\item{plotting}{A boolean value to determine whether to plot the scree plot or not, default to be False}

\item{proj_dataset}{The datasets to be projected on}

\item{proj_group}{The grouping of projected data sets}

\item{enable_normalization}{An argument to decide whether to use normalizaiton or not,  default is TRUE}

\item{column_sum_normalization}{An argument to decide whether to use column sum normalization or not, default it TRUE}

\item{screen_prob}{A vector of probabilies for genes to be chosen}
}
\value{
A list contains the component and the score of each dataset on every component after twoStageiLCA.rank algorithm
}
\description{
Two-staged decomposition of several matrices with Independent LCA,
twoStageLCA is first performed on the data,
the rank selection procedure is automatic based on BEMA.
Then, fastICA is implemented on the score to extract the independent
components.
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1, 2, 3, 4), c(1, 2), c(3, 4), c(1, 3), c(2, 4), c(1), c(2), c(3), c(4))
threshold = c(3, 1.5, 1.5, 1.5, 1.5, 0.5, 0.5, 0.5, 0.5)
res_twoStageiLCA.rank = twoStageiLCA.rank(
dataset,
group,
threshold = threshold)

}
\keyword{LCA}
\keyword{indepentdent,}
\keyword{rank,}
\keyword{two-staged,}
