% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageLCA.R
\name{twoStageLCA}
\alias{twoStageLCA}
\title{Two-staged Linked Component Analysis}
\usage{
twoStageLCA(
  dataset,
  group,
  comp_num,
  weighting = NULL,
  backup = 0,
  plotting = FALSE,
  proj_dataset = NULL,
  proj_group = NULL,
  enable_normalization = TRUE,
  column_sum_normalization = TRUE,
  screen_prob = NULL
)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{weighting}{Weighting of each dataset, initialized to be NULL}

\item{backup}{A positive scalar to determine how many PCs to over select}

\item{plotting}{A boolean value to determine whether to plot the scree plot or not, default to be False}

\item{proj_dataset}{The datasets to be projected on}

\item{proj_group}{The grouping of projected data sets}

\item{enable_normalization}{An argument to decide whether to use normalizaiton or not,  default is TRUE}

\item{column_sum_normalization}{An argument to decide whether to use column sum normalization or not, default it TRUE}

\item{screen_prob}{A vector of probabilies for genes to be chosen}
}
\value{
A list contains the component and the score of each dataset on every component after 2sLCA algorithm
}
\description{
Two-staged Linked Component Analysis
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1, 2, 3, 4), c(1, 2), c(3, 4), c(1, 3), c(2, 4), c(1), c(2), c(3), c(4))
comp_num = c(2, 2, 2, 2, 2, 2, 2, 2, 2)
proj_dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
proj_group = list(c(TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE))
res_twoStageLCA = twoStageLCA(
dataset,
group,
comp_num,
proj_dataset = proj_dataset,
proj_group = proj_group)

}
\keyword{LCA}
\keyword{two-staged,}
