% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_tsne_onLC.R
\name{umap_tsne_onLC}
\alias{umap_tsne_onLC}
\title{genetrate new UMAP and TSNE coordinates given LCA object}
\usage{
umap_tsne_onLC(twoStageLCAobj, group, n_comp, add_to_meta, meta_list)
}
\arguments{
\item{twoStageLCAobj}{list output of tsLCA run, list}

\item{group}{name of the weights group from the SJD_algorithm output, i.e 'Shared.All.13'}

\item{n_comp}{number of components to make umaps and tsne, int}

\item{add_to_meta}{adds tsne and umap coordinates to metadata, bool}

\item{meta_list}{Meta.List that was passed into SJD run, str}
}
\description{
genetrate new UMAP and TSNE coordinates given LCA object
}
