% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatch.R
\name{getMatch}
\alias{getMatch}
\title{Species matching function}
\usage{
getMatch(
  genes,
  inSpecies,
  inType,
  newSpecies,
  moreAttrIn = NA,
  moreAttrNew = NA
)
}
\arguments{
\item{genes}{vector of gene names in either ensembl or symbol format}

\item{inSpecies}{character of species name of input genes i.e 'human'}

\item{inType}{character of type of input genes i.e 'ensembl'}

\item{newSpecies}{character of target species for the new vector of genes i.e 'mouse'}

\item{moreAttrIn}{character of other gene attributes i.e 'species latin'}

\item{moreAttrNew}{character of other gene attributes i.e 'ensembl.nms'}
}
\value{
vector of matched genes from the target species
}
\description{
matching genes from one species (i.e human) to another species (i.e mouse) and output genes
}
\examples{

data(NeuroGenesis4)
out = getMatch(
rownames(NeuroGenesis4$Meissner.inVitro.bulk.Hs),
inSpecies = 'human',
inType = 'symbol',
newSpecies = 'mouse')

}
\keyword{genes}
\keyword{shared}
