% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatch.R
\name{getMatch}
\alias{getMatch}
\title{Get Gene Match}
\usage{
getMatch(
  genes,
  inSpecies,
  inType,
  newSpecies,
  moreAttrIn = NA,
  moreAttrNew = NA
)
}
\arguments{
\item{genes}{A vector of gene}

\item{inSpecies}{Kind of species}

\item{inType}{Kind of gene type}

\item{newSpecies}{Output species}

\item{moreAttrIn}{More attributes input}

\item{moreAttrNew}{More attributes output}
}
\value{
A vector of genes
}
\description{
Get genes match for a single input data set
}
\keyword{gene,}
\keyword{match}
