% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatch.R
\name{getMatch}
\alias{getMatch}
\title{Species matching function}
\usage{
getMatch(
  genes,
  inSpecies,
  inType,
  newSpecies,
  useNewestVersion = FALSE,
  moreAttrIn = NA,
  moreAttrNew = NA
)
}
\arguments{
\item{genes}{character vector of gene identifiers in either ensembl gene identifier or gene symbol format.}

\item{inSpecies}{a character indicating the species from which 'genes' come; must be one of c("human","mouse","roundworm","fruitfly","zebrafish","chicken","rat","guinea pig","golden hamster","rabbit","pig","sheep","cow","dog","cat","macaque","bonobo","chimpanzee")}

\item{inType}{inType: single character of type of input genes, must be one of c("symbol",'ensembl'")}

\item{newSpecies}{a character indicating the species for which orthologous gene identifiers are desired; must be one of c("human","mouse","roundworm","fruitfly","zebrafish","chicken","rat","guinea pig","golden hamster","rabbit","pig","sheep","cow","dog","cat","macaque","bonobo","chimpanzee")}

\item{useNewestVersion}{logical indicating if the function should attempt to use the latest version of ensembl, or to use the Aug 2020 archive version that is more stable. default is FALSE.}

\item{moreAttrIn}{character vector of other gene attributes that you want returned for the input species from biomaRt - to add this argument you must know the names of the fields in the species-specific biomaRt that you are requesting. default is NA.}

\item{moreAttrNew}{character vector of other gene attributes that you want returned for the output species from biomaRt - to add this argument you must know the names of the fields in the species-specific biomaRt that you are requesting. default is NA.}
}
\value{
A matrix whose first column contains the exact gene identifiers (and in the same order) that were sent to the function in the "genes" argument, followed by 3 columns of gene identifiers that were retrieved from biomaRt from both the "inSpecies" and the "newSpecies" (for each species, these 3 identifiers are: gene symbol, ensembl gene ID, and text description).
}
\description{
Use biomaRt to retrieve additional gene identifiers for a vector of gene identifiers in one species (e.g. 'human'), or gene identifiers for orthologous genes in a second species (e.g. 'mouse')
}
\examples{

data(NeuroGenesis4)
out = getMatch(
rownames(NeuroGenesis4$Meissner.inVitro.bulk.Hs),
inSpecies = 'human',
inType = 'symbol',
newSpecies = 'mouse')

}
\keyword{genes}
\keyword{shared}
