% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeData.R
\name{normalizeData}
\alias{normalizeData}
\title{Data Normalization}
\usage{
normalizeData(dataset, enable_normalization = TRUE)
}
\arguments{
\item{dataset}{The input list of data sets matrix}

\item{enable_normalization}{An argument to decide whether to use normalizaiton or not,  default is TRUE}
}
\description{
Normalize data to have mean zero and std 1
}
\examples{
dataset = list(
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50)
)
normalizeData(dataset, TRUE)

}
\keyword{normalize}
