% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectNMF.R
\name{projectNMF}
\alias{projectNMF}
\title{Projection function for Joint Decomposition with Nonnegative Matrix Factorization
Purpose: finding out the common development patterns between new data and datasets used in the joint decomposition.
If you have new dataset(s) and Gene scores produced by JointNMF, this function can produce estimated sample scores for the new datasets.}
\usage{
projectNMF(
  proj_dataset,
  proj_group,
  list_component,
  max_ite = 1000,
  max_err = 1e-04,
  enable_normalization = TRUE,
  column_sum_normalization = FALSE
)
}
\arguments{
\item{proj_dataset}{The dataset(s) to be projected on.}

\item{proj_group}{A listed of boolean combinations indicating which groupings should be used for each projected dataset.The length of proj_group should match the length of proj_dataset, and the length of each concatenated boolean combination should match the length of the parameter group.}

\item{list_component}{list_component produced from JointNMF decomposition.}

\item{max_ite}{The maximum number of iterations for the jointNMF algorithms to run, default value is set to 100}

\item{max_err}{The maximum error of loss between two iterations, or the program will terminate and return, default value is set to be 0.0001}

\item{enable_normalization}{An argument to decide whether to use normalizaiton or not,  default is TRUE}

\item{column_sum_normalization}{An argument to decide whether to use column sum normalization or not, default it FALSE}
}
\value{
A list contains the projected scores of each dataset on every component.
}
\description{
Projection function for Joint Decomposition with Nonnegative Matrix Factorization
Purpose: finding out the common development patterns between new data and datasets used in the joint decomposition.
If you have new dataset(s) and Gene scores produced by JointNMF, this function can produce estimated sample scores for the new datasets.
}
\examples{
proj_dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
proj_group = list(c(TRUE, TRUE)) # which groupings in the joint decomposition you want to project on.
list_component = jointNMF$linked_component_list # from jointNMF result
res_projNMF = projectNMF(
proj_dataset = proj_dataset,
proj_group = proj_group,
list_component = list_component)

PLEASE MAKE SURE YOUR proj_dataset AND list_component ELEMENTS HAVE MEANINGFUL ROW(GENE) NAMES - they are matched across matrices for the projection. 
#'
}
\keyword{NMF}
\keyword{joint,}
\keyword{projection,}
