% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAssignment.R
\name{scoreNameAssignProj}
\alias{scoreNameAssignProj}
\title{Score Name Assignment for projectNMF}
\usage{
scoreNameAssignProj(score_list, group_name)
}
\arguments{
\item{score_list}{A list of scores in the analysis}

\item{group_name}{A vector of group names extracted by groupNameExtractor function}
}
\value{
A list of scores assigned with names
}
\description{
Assign names to score lists based on group_name,
the group_name
is the output of the groupNameExtractor
}
\examples{
score_list = list(
list(matrix(c(1 : 4), nrow = 2), matrix(c(1 : 4), nrow = 2)),
list(matrix(c(1 : 4), nrow = 2), matrix(c(1 : 4), nrow = 2)))
group_name = c("comp1", "comp2")
scoreNameAssignProj(score_list, group_name)

}
\keyword{name}
\keyword{score,}
