% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepICA.R
\name{sepICA}
\alias{sepICA}
\title{Single Data Set Decomposition with Independent Component Analysis}
\usage{
sepICA(dataset, comp_num, weighting = NULL)
}
\arguments{
\item{dataset}{A dataframe/matrix to be decomposed}

\item{comp_num}{Number of ICs to be extracted}

\item{weighting}{Weighting of each dataset, initialized to be NULL}
}
\value{
A list of scores and component
}
\description{
Apply ICA (Independent Component Analysis) to a single data set
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepICA = sepICA(dataset, comp_num)

}
\keyword{ICA}
\keyword{analysis,}
\keyword{separate}
